x64: replace uint32_t with uintptr_t for host pointers in dynarec

+++ src/Core/R4300/x86/GCop0.cpp
@@ -8,22 +8,22 @@
 #include <R4300/Ops.hpp>
 #include <R4300/Recomph.hpp>
 
-// static uint32_t pMFC0 = (uint32_t)(MFC0);
+// static uint32_t pMFC0 = (uintptr_t)(MFC0);
 void genmfc0()
 {
-    gencallinterp((uint32_t)MFC0, 0);
+    gencallinterp((uintptr_t)MFC0, 0);
     /*dst->local_addr = code_length;
-    mov_m32_imm32((void *)(&PC), (uint32_t)(dst));
+    mov_m32_imm32((void *)(&PC), (uintptr_t)(dst));
     call_m32((uint32_t *)(&pMFC0));
     genupdate_system(0);*/
 }
 
-// static uint32_t pMTC0 = (uint32_t)(MTC0);
+// static uint32_t pMTC0 = (uintptr_t)(MTC0);
 void genmtc0()
 {
-    gencallinterp((uint32_t)MTC0, 0);
+    gencallinterp((uintptr_t)MTC0, 0);
     /*dst->local_addr = code_length;
-    mov_m32_imm32((void *)(&PC), (uint32_t)(dst));
+    mov_m32_imm32((void *)(&PC), (uintptr_t)(dst));
     call_m32((uint32_t *)(&pMTC0));
     genupdate_system(0);*/
 }
